fix the course number 24 bug :|

+++ app/Http/Controllers/ThreadsController.php
@@ -31,10 +31,9 @@ class ThreadsController extends Controller
     {
 
 
-        $threads = $this->getThreads($channel,$filters);
+        $threads = $this->getThreads($channel, $filters);
 
-        if(request()->wantsJson())
-        {
+        if (request()->wantsJson()) {
             return $threads;
         }
 
@@ -123,12 +122,17 @@ class ThreadsController extends Controller
      * @param  \App\Thread  $thread
      * @return \Illuminate\Http\Response
      */
-    public function destroy($channel,Thread $thread)
+    public function destroy($channel, Thread $thread)
     {
+
+        $this->authorize('update',$thread);
+
+
         $thread->delete();
 
-        if(request()->wantsJson())
-            return response([],Response::HTTP_NO_CONTENT);
+        if (request()->wantsJson())
+            return response([], 204);
+        
 
         return redirect('/threads');
     }
@@ -144,6 +148,5 @@ class ThreadsController extends Controller
         }
 
         return $threads->get();
-
     }
 }
